% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-brush-xspline.R, R/geom-offset-bezier.R,
%   R/geom-offset-xspline.R, R/geom-vwarc.R, R/geom-vwcurve.R,
%   R/geom-vwdiagonal.R, R/geom-vwline.R, R/geom-vwxspline.R, R/geom_vwzezier.R,
%   R/ggproto.R
\docType{data}
\name{GeomBrushXspline}
\alias{GeomBrushXspline}
\alias{GeomOffsetBezier}
\alias{GeomOffsetXspline}
\alias{StatVwarc}
\alias{GeomVwcurve}
\alias{StatVwdiagonal}
\alias{GeomVwline}
\alias{GeomVwXspline}
\alias{StatVwbezier}
\alias{ggvwline-extensions}
\title{ggvwline extensions to ggplot2}
\format{
An object of class \code{StatVwdiagonal} (inherits from \code{StatDiagonal}, \code{Stat}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
StatVwdiagonal
}
\description{
In general the actual classes should be of little interest to users as the
standard ggplot2 api of using geom_* and stat_* functions for building up
the plot is encouraged.
}
\keyword{datasets}
