% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logis-curve.R
\name{logisCurveGrob}
\alias{logisCurveGrob}
\alias{grid.logis_curve}
\title{Draw Logistic Curve
These functions create and draw logistic curve.}
\usage{
logisCurveGrob(
  x = 0,
  y = 0,
  xend = 1,
  yend = 1,
  a = 0,
  b = 0.05,
  k = 20,
  n = 200,
  ...
)

grid.logis_curve(...)
}
\arguments{
\item{x}{a numeric vector specifying the x-location of the start point.}

\item{y}{a numeric vector specifying the y-location of the start point.}

\item{xend}{a numeric vector specifying the x-location of the end point.}

\item{yend}{a numeric vector specifying the y-location of the end point.}

\item{a}{a numeric vector specifying the center point.}

\item{b}{a numeric vector specifying the growth rate.}

\item{k}{a numeric vector specifying the initial value.}

\item{n}{positive integer value.}

\item{...}{others parameters.}
}
\description{
Draw Logistic Curve
These functions create and draw logistic curve.
}
