% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{verticalBrush}
\alias{circleBrush}
\alias{squareBrush}
\alias{widthSpec}
\alias{widthSpline}
\alias{BezierWidth}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{vwline}{\code{\link[vwline:widthSpec]{BezierWidth}}, \code{\link[vwline:grid.brushXspline]{circleBrush}}, \code{\link[vwline:grid.brushXspline]{squareBrush}}, \code{\link[vwline:grid.brushXspline]{verticalBrush}}, \code{\link[vwline]{widthSpec}}, \code{\link[vwline:widthSpec]{widthSpline}}}
}}

